[Refactor][AscendStore] Move the layerwise protocol into the memcache backend
make_full_key / make_partial_key / make_hit_check_keys / extract_layout_config now live in memcache_backend.py as neutral module-level functions; gva_protocol.py becomes a backward-compatible alias (GVAKeyFactory delegates to the moved functions) so consumers and tests keep working until they switch over. Key strings are byte-identical to the pre-move formats.
